Added a register client acknowledgement packet.
It generates a random userid to tie an loose auth mechanism to.
This prevents players from sabotaging and cheating.  We are presently
pulling off the /dev/urandom file, which is a non-blocking entropy
pool supported by many platforms.  In Solaris with SPARC, it will
make use of the KCF, the Kernel Cryptographic Framework, to pull
out cryptographically secure random numbers using hardware RNG
on T2 or greater platforms.  In the even this is compiled for a
platform that doesn't have a /dev/urandom device, I've given an
option for specifying the entropy file location (e.g. I think FreeBSD
only has /dev/random, though on newer versions a symlink exists to it).
random(3C) on Solaris specifies developers should use /dev/urandom for
crypto-secure random.  Because being a CSRNG isn't really that important,
this may be modified in the future to make the idler loop every so often
reseed a simpler PRNG.  This prevents us from having make syscalls every
time we need a random number (i.e. read(2)).