Added a length field to the kick message.
This is somewhat necessary to know whether the client is receiving
the message in its entirety or just receiving a partial string.  In
any case the client will disconnect, as the client cannot request
the packet again (otherwise the server would be asking for a UDP storm).
The error handling in this case will likely be the client warning the
user and then exiting with the partial message.