Added more validation for the packets.
We now have something to constrain room creation based on name
length and printable characters. Removed REG_ACK messages from
the invalid server receive side case, as it's now expected behavior
that the client replies back to the same message to validate that they
are in fact a valid participant.