Moved the location of rwlock rdunlock.
This is important, as the rwlock is protecting more than just
the hashtable, but also accesses to the allocated players as well.
If, for instance, the main loop decides to expire or timeout a user,
thereby freeing them, then the code could potentially end up with
a use after free bug.  By holding onto the read lock, a user cannot
be expired killed, kicked, or destroyed while asynchronous calls
are handled in callbacks.