Tidied things up quite a bit.
I am considering moving the packet processing logic to packet.c,
as this was initially in tetris_serv.c in an attempt to keep all
callbacks and uv_main_loop-related things together.  However,
a tetris_serv.c file that contains every variable and function
ever used in the uv loop is a bit cumbersome and the file is
becoming so large it's difficult to navigate.