Made things compile cleanly on Linux.
Had to invert the extern/declaration relationship for the hashtables,
as evidently gcc will emit a compiler warning if you redeclare an
extern.  It probably makes way more sense to add a global server state
object somewhere and use that to store all the global server-like things
somewhere, but for now we'll not do that.