Fixed a bug with freeing stack memory.
lsrooms will now show the status of the rooms in the list as well.
When a packet was being freed, an array inside it was also being
freed, this is technically illegal, despite the actual allocation
happening on the heap, free is actually treating the char[] as a
contiguous array of bytes with the struct.  Added the player who
created the room to the room list.  There will need to be game state
tracking logic for each room that will correctly handle when the player
list drops below 1 in length.