Have random public ids now.
This field is meant to prevent cheating, as the player doesn't
have knowledge of the private id of the opponent (used in part for authorizing
packets).

I tried to take special care with the locking, but it's still possible there
will be some bugs here.  It will be very tricky to get right so things don't
induce either a.) a deadlock, or b.) a use after free bug.  The assigning of
a field in the player_t struct in particular seems a bit worrisome, we'll have
to be sure that a player cannot be destroyed without first disowning the player
from the room, but the room's list of players cannot be modified without first
obtaining the room lock.  We may need to use our giant global locks as placeholders
if we encounter bugs here.  It will be worth fuzzing with packets as best we can to
attempt to induce some deadlocking / crashing. In particular trying to send disconnect
packets at the same time the room is operating on the player.