Have room destruction working.
However the handling of the reader writer locks is super goofy,
we probably need to establish a better precendence for how these
locks are resolved and who destroys whom.  We need to make sure a
user can't destroy themselves at the same time a room is trying to
disconnect the user, otherwise we'll either hit a deadlock or use
after free bug.