Made this player list more sensible.
Sometimes a linked list is not the easiest of solutions for a tiny
enough list.  Locking and other operations have been made a bit simpler
without being constrained to g_slist's foreach function signature, as well
as the fact that we can traverse the entire list forward and backward with
the assumption that we have a bounded size that is reasonable (MAX_PLAYERS = 4).