Storing game states within the room.
Also starting to break up the monolithic tetris.h and tetris.c
elements to be more modular and reusable.  I'm using some macros
to abuse the shared types for now (a good portion of the game state
is actually useless to the server, as it deals with client-side config
stuff).  This may more appropriately belong elsewhere.